Add dependencies, add build configuration and restructure project packages.
