+++ .scalafmt.conf
@@ -1,0 +1,44 @@
+version = 2.6.4
+
+// Sets max amount of characters before new line
+maxColumn = 160
+
+// Sets indentation of method arguments and parameters
+continuationIndent {
+  callSite = 2
+  defnSite = 2
+}
+
+// Aligns tokens for numerous methods e.g. matchers, for yield and module ID's
+align = more
+
+
+// PreferCurlyFors        - Replaces parentheses into curly braces in for comprehensions that contain multiple enumerator generators.
+// RedundantBraces        - Removes redundant brances
+// RedundantParens        - Removes reduanant parentheses, including those within string interpolation
+// SortModifiers          - Sorts modifiers alphabetically
+// ExpandImportSelectors  - Adds collapsed imports into new imports
+rewrite {
+  rules = [PreferCurlyFors, RedundantBraces, RedundantParens, SortModifiers, ExpandImportSelectors]
+  redundantBraces {
+    stringInterpolation = true
+  }
+}
+
+newlines {
+  alwaysBeforeTopLevelStatements = true
+  alwaysBeforeCurlyBraceLambdaParams = true
+}
+
+// Adds breaks between large lists
+binPack.literalArgumentLists = true
+
+includeCurlyBraceInSelectChains = true
+optIn.breakChainOnFirstMethodDot = true
+
+rewriteTokens {
+  "\t": "  "
+  "→": "->"
+  "←": "<-"
+  "⇒": "=>"
+}
