+++ app/config/AppConfig.scala
@@ -14,20 +14,18 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.customsreferencedata.controllers
+package config
 
-import javax.inject.{Inject, Singleton}
-import play.api.mvc.{Action, AnyContent, ControllerComponents}
-import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
-import uk.gov.hmrc.customsreferencedata.config.AppConfig
+import javax.inject.Inject
+import javax.inject.Singleton
+import play.api.Configuration
+import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
 
-import scala.concurrent.Future
+@Singleton
+class AppConfig @Inject() (config: Configuration, servicesConfig: ServicesConfig) {
 
-@Singleton()
-class MicroserviceHelloWorldController @Inject()(appConfig: AppConfig, cc: ControllerComponents)
-    extends BackendController(cc) {
+  val authBaseUrl: String = servicesConfig.baseUrl("auth")
 
-  def hello(): Action[AnyContent] = Action.async { implicit request =>
-    Future.successful(Ok("Hello world"))
-  }
+  val auditingEnabled: Boolean = config.get[Boolean]("auditing.enabled")
+  val graphiteHost: String     = config.get[String]("microservice.metrics.graphite.host")
 }
