+++ conf/application.conf
@@ -25,6 +25,10 @@ play.application.loader = "uk.gov.hmrc.play.bootstrap.ApplicationLoader"
 # Primary entry point for all HTTP requests on Play applications
 play.http.requestHandler = "uk.gov.hmrc.play.bootstrap.http.RequestHandler"
 
+
+# Application module for buisness logic modules
+play.modules.enabled += "config.Module"
+
 # Provides an implementation of AuditConnector. Use `uk.gov.hmrc.play.bootstrap.AuditModule` or create your own.
 # An audit connector must be provided.
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.AuditModule"
@@ -47,7 +51,7 @@ play.http.errorHandler = "uk.gov.hmrc.play.bootstrap.backend.http.JsonErrorHandl
 # ~~~~
 # Additional play modules can be added here
 
-play.modules.enabled += "play.modules.reactivemongo.ReactiveMongoHmrcModule"
+ play.modules.enabled += "play.modules.reactivemongo.ReactiveMongoModule"
 
 # Secret key
 # ~~~~~
@@ -81,7 +85,7 @@ controllers {
   # 300 is the default, you may need to change this according to your needs
   confidenceLevel = 300
 
-  uk.gov.hmrc.customsreferencedata.controllers.MicroserviceHelloWorldController = {
+  controllers.MicroserviceHelloWorldController = {
     needsAuth = false
     needsLogging = false
     needsAuditing = false
