+++ project/plugins.sbt
@@ -17,3 +17,5 @@ addSbtPlugin("uk.gov.hmrc" % "sbt-distributables" % "2.0.0")
 addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.7.5")
 
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.6.0")
+
+addSbtPlugin("org.scalameta" % "sbt-scalafmt" % "2.4.2")
